requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
google-api-python-client
Pillow